(+) generate_table1();  separated pyscripts; (+) unittests
