+++ src/eda_toolkit/art.py
@@ -5,14 +5,14 @@
 import os
 
 eda_toolkit_logo = r""" 
-+--------------------------------------------------------------------------------+
-|  _____   ____       _              _____                   _   _      _   _    |
-| | ____| |  _ \     / \            |_   _|   ___     ___   | | | | __ (_) | |_  |
-| |  _|   | | | |   / _ \    _____    | |    / _ \   / _ \  | | | |/ / | | | __| |
-| | |___  | |_| |  / ___ \  |_____|   | |   | (_) | | (_) | | | |   <  | | | |_  |
-| |_____| |____/  /_/   \_\           |_|    \___/   \___/  |_| |_|\_\ |_|  \__| |
-|                                                                                |
-+--------------------------------------------------------------------------------+                                                                                                                                                            
+
+███████╗██████╗  █████╗       ████████╗ ██████╗  ██████╗ ██╗     ██╗  ██╗██╗████████╗
+██╔════╝██╔══██╗██╔══██╗      ╚══██╔══╝██╔═══██╗██╔═══██╗██║     ██║ ██╔╝██║╚══██╔══╝
+█████╗  ██║  ██║███████║█████╗   ██║   ██║   ██║██║   ██║██║     █████╔╝ ██║   ██║   
+██╔══╝  ██║  ██║██╔══██║╚════╝   ██║   ██║   ██║██║   ██║██║     ██╔═██╗ ██║   ██║   
+███████╗██████╔╝██║  ██║         ██║   ╚██████╔╝╚██████╔╝███████╗██║  ██╗██║   ██║   
+╚══════╝╚═════╝ ╚═╝  ╚═╝         ╚═╝    ╚═════╝  ╚═════╝ ╚══════╝╚═╝  ╚═╝╚═╝   ╚═╝   
+                                                                                                                                                                                                                                            
 """
 
 ###################################################################################
