+++ unittests/test_plots.py
@@ -1,38 +1,24 @@
+# plotting_tests.py
+
 import pytest
-import importlib
+import os
+import io
+import sys
+from unittest.mock import patch
 import matplotlib.pyplot as plt
 import pandas as pd
 import numpy as np
-import io
-import sys
-import os
-import datetime
-import builtins
-from scipy import stats
-from unittest import mock
-from unittest.mock import MagicMock, patch
 from eda_toolkit import (
-    ensure_directory,
-    add_ids,
-    strip_trailing_period,
-    parse_date_with_rule,
-    dataframe_profiler,
-    contingency_table,
-    save_dataframes_to_excel,
-    summarize_all_combinations,
     kde_distributions,
     data_doctor,
     flex_corr_matrix,
     scatter_fit_plot,
     box_violin_plot,
     stacked_crosstab_plot,
-    highlight_columns,
-    custom_help,
-    eda_toolkit_logo,
-    detailed_doc,
 )
 
 
+# Fixtures that relate to data used in plotting
 @pytest.fixture
 def sample_dataframe():
     return pd.DataFrame({"Feature1": [1, 2, 3, 4, 5], "Feature2": [10, 20, 30, 40, 50]})
@@ -41,47 +27,6 @@ def sample_dataframe():
 @pytest.fixture
 def sample_dataframe_values():
     return pd.DataFrame({"values": [1, 2, 3, 4, 5]})
-
-
-@pytest.fixture
-def sample_dataframes():
-    return {
-        "Sheet1": pd.DataFrame({"A": [1, 2, 3], "B": [4, 5, 6]}),
-        "Sheet2": pd.DataFrame({"X": [7, 8, 9], "Y": [10, 11, 12]}),
-    }
-
-
-@pytest.fixture
-def sample_dataframe_with_text():
-    return pd.DataFrame({"col": ["123.", "456", "789.", None, "text."]})
-
-
-@pytest.fixture
-def sample_dataframe_dates():
-    return pd.DataFrame({"dates": ["25/12/2023", "12/25/2023", "05/06/2023"]})
-
-
-@pytest.fixture
-def sample_dataframe_profiler():
-    return pd.DataFrame({"col1": [1, 2, 2, None], "col2": ["A", "B", "B", "A"]})
-
-
-@pytest.fixture
-def sample_dataframe_contingency():
-    return pd.DataFrame(
-        {"Category": ["A", "B", "A", "B", "C"], "Values": [1, 2, 1, 2, 3]}
-    )
-
-
-@pytest.fixture
-def sample_categorical_dataframe():
-    return pd.DataFrame(
-        {
-            "Category": ["A", "A", "B", "B", "C"],
-            "Subcategory": ["X", "Y", "X", "Y", "X"],
-            "Value": [10, 15, 20, 25, 30],
-        }
-    )
 
 
 @pytest.fixture
@@ -97,7 +42,6 @@ def sample_corr_dataframe():
 
 @pytest.fixture
 def sample_scatter_dataframe():
-    """Create a sample DataFrame for scatter plots."""
     np.random.seed(42)
     return pd.DataFrame(
         {
@@ -106,6 +50,18 @@ def sample_scatter_dataframe():
             "Category": np.random.choice(["A", "B"], size=50),
         }
     )
+
+
+@pytest.fixture
+def sample_boxcox_dataframe():
+    """Fixture to provide test data for Box-Cox transformation."""
+    return pd.DataFrame({"values": np.array([1, 2, 3, 4, 5])})
+
+
+@pytest.fixture
+def sample_boxcox_invalid_dataframe():
+    """Fixture with non-positive values, which should trigger a ValueError."""
+    return pd.DataFrame({"values": np.array([-1, 0, 2, 3, 4])})
 
 
 @pytest.fixture
@@ -132,165 +88,10 @@ def sample_crosstab_dataframe():
     )
 
 
-@pytest.fixture
-def sample_boxcox_dataframe():
-    """Fixture to provide test data for Box-Cox transformation."""
-    return pd.DataFrame({"values": np.array([1, 2, 3, 4, 5])})
-
-
-@pytest.fixture
-def sample_boxcox_invalid_dataframe():
-    """Fixture with non-positive values, which should trigger a ValueError."""
-    return pd.DataFrame({"values": np.array([-1, 0, 2, 3, 4])})
-
-
-def test_ensure_directory(tmp_path):
-    dir_path = tmp_path / "test_dir"
-    ensure_directory(str(dir_path))
-    assert dir_path.exists()
-
-
-def test_save_dataframes_to_excel(sample_dataframes, tmp_path):
-    file_path = tmp_path / "test_output.xlsx"
-
-    try:
-        # Debugging
-        print(f"DEBUG: file_path = {file_path}, type = {type(file_path)}")
-        assert isinstance(
-            file_path, (str, os.PathLike)
-        ), "file_path must be a string or Path-like object"
-
-        save_dataframes_to_excel(
-            file_path,  # Ensure file path is first argument
-            sample_dataframes,  # Ensure dataframes dictionary is passed correctly
-        )
-
-        assert file_path.exists()  # Ensure the file was created
-    except Exception as e:
-        pytest.fail(f"save_dataframes_to_excel failed: {e}")
-
-
-def test_add_ids(sample_dataframe):
-    num_digits = 5
-    seed = 42
-
-    df_with_ids = add_ids(
-        sample_dataframe, id_colname="UniqueID", num_digits=num_digits, seed=seed
-    )
-
-    # Check if UniqueID column is present
-    assert "UniqueID" in df_with_ids.columns, "Column 'UniqueID' was not added"
-
-    # Check that all UniqueID values are unique
-    assert df_with_ids["UniqueID"].nunique() == len(
-        sample_dataframe
-    ), "Duplicate IDs found"
-
-    # Check that the length of each ID matches num_digits
-    assert (
-        df_with_ids["UniqueID"].apply(lambda x: len(str(x)) == num_digits).all()
-    ), f"Some IDs do not have {num_digits} digits"
-
-    # Check if the function is deterministic (same seed should generate same IDs)
-    df_with_ids_2 = add_ids(
-        sample_dataframe, id_colname="UniqueID", num_digits=num_digits, seed=seed
-    )
-
-    assert (
-        df_with_ids["UniqueID"].tolist() == df_with_ids_2["UniqueID"].tolist()
-    ), "IDs generated with the same seed do not match"
-
-
-def test_add_ids_exceed_pool_size():
-    """Test that a ValueError is raised when the number of rows exceeds the pool of unique IDs."""
-
-    # Set num_digits such that the pool size is too small for the given DataFrame
-    num_digits = 3  # Pool size = 9 * (10^2) = 900 unique IDs
-    pool_size = 9 * (10 ** (num_digits - 1))  # Pool size = 900
-    n_rows = pool_size + 1  # Exceed the pool size
-
-    # Create a DataFrame with more rows than the pool size
-    df = pd.DataFrame({"data": range(n_rows)})
-
-    # Expect a ValueError when trying to add IDs
-    with pytest.raises(
-        ValueError,
-        match=r"The number of rows \(\d+\) exceeds the total pool of possible unique IDs",
-    ):
-        add_ids(df, num_digits=num_digits)
-
-
-def test_strip_trailing_period(sample_dataframe_with_text):
-    """Test that strip_trailing_period removes trailing periods correctly."""
-
-    df_cleaned = strip_trailing_period(sample_dataframe_with_text, "col")
-
-    # Define expected results (ensure None values match actual Pandas NaN behavior)
-    expected_values = ["123", "456", "789", np.nan, "text"]
-
-    # Convert column to list and check each value (handling NaN cases)
-    cleaned_values = df_cleaned["col"].tolist()
-    for actual, expected in zip(cleaned_values, expected_values):
-        if pd.isna(expected):
-            assert pd.isna(actual), f"Expected NaN but got {actual}"
-        else:
-            assert actual == expected, f"Expected {expected} but got {actual}"
-
-    # Ensure the function does not modify other columns
-    df_original = sample_dataframe_with_text.copy()
-    df_cleaned_other = strip_trailing_period(df_original, "col")
-
-    assert df_cleaned_other.drop(columns=["col"]).equals(
-        df_original.drop(columns=["col"])
-    ), "Function should not modify other columns"
-
-    # Ensure type consistency (all values should remain str or NaN)
-    assert all(
-        isinstance(val, (str, float)) or pd.isna(val) for val in df_cleaned["col"]
-    ), "All values should remain as strings or NaN"
-
-
-def test_parse_date_with_rule(sample_dataframe_dates):
-    assert parse_date_with_rule(sample_dataframe_dates.iloc[0, 0]) == "2023-12-25"
-    assert parse_date_with_rule(sample_dataframe_dates.iloc[1, 0]) == "2023-12-25"
-    assert parse_date_with_rule(sample_dataframe_dates.iloc[2, 0]) == "2023-06-05"
-
-
-def test_dataframe_profiler(sample_dataframe_profiler):
-    profile = dataframe_profiler(sample_dataframe_profiler, return_df=True)
-    assert "column" in profile.columns
-    assert "null_total" in profile.columns
-
-
-def test_contingency_table(sample_dataframe_contingency):
-    table = contingency_table(sample_dataframe_contingency, cols="Category")
-    assert "Total" in table.columns
-    assert "Percentage" in table.columns
-
-
-def test_summarize_all_combinations(sample_categorical_dataframe, tmp_path):
-    try:
-        data_path = str(tmp_path)
-        data_name = "summary.xlsx"
-        summary_tables, all_combinations = summarize_all_combinations(
-            df=sample_categorical_dataframe,
-            variables=["Category", "Subcategory"],
-            data_path=data_path,
-            data_name=data_name,
-            min_length=2,
-        )
-
-        assert isinstance(summary_tables, dict), "Output should be a dictionary"
-        assert all(
-            isinstance(df, pd.DataFrame) for df in summary_tables.values()
-        ), "Each summary should be a DataFrame"
-        assert all_combinations, "All combinations list should not be empty"
-        assert os.path.exists(
-            os.path.join(data_path, data_name)
-        ), "Excel file should be created"
-
-    except Exception as e:
-        pytest.fail(f"summarize_all_combinations failed: {e}")
+@pytest.fixture(autouse=True, scope="session")
+def suppress_plot_show():
+    """Automatically mock plt.show() so figures don't pop up during tests."""
+    plt.show = lambda *args, **kwargs: None
 
 
 @pytest.mark.parametrize(
@@ -304,6 +105,8 @@ def test_summarize_all_combinations(sample_categorical_dataframe, tmp_path):
         ("kde", None),
     ],
 )
+
+# < Insert all test_ functions for kde_distributions, data_doctor, scatter_fit_plot, box_violin_plot, flex_corr_matrix, stacked_crosstab_plot >
 def test_kde_distributions(sample_dataframe, tmp_path, plot_type, vars_of_interest):
     """Test kde_distributions with different plot types and variable selections."""
     save_path = str(tmp_path)
@@ -366,6 +169,140 @@ def test_kde_distributions_log_scale(sample_dataframe, tmp_path):
         ), "PNG file should be created for log scale."
     except Exception as e:
         pytest.fail(f"kde_distributions with log scale failed: {e}")
+
+
+def test_kde_distributions_invalid_std_color_length(sample_dataframe):
+    with pytest.raises(
+        ValueError,
+        match="Not enough colors specified in 'std_color'",
+    ):
+        kde_distributions(
+            df=sample_dataframe,
+            vars_of_interest=["Feature1"],
+            plot_type="hist",
+            plot_mean=True,
+            plot_median=True,
+            std_dev_levels=[1, 2],
+            std_color=["#888888"],  # mismatch here
+            show_legend=False,
+        )
+
+
+def test_kde_distributions_single_var_grid_warning(sample_dataframe):
+    with pytest.raises(
+        ValueError,
+        match="Cannot use grid_figsize when there is only one",
+    ):
+        kde_distributions(
+            df=sample_dataframe,
+            vars_of_interest=["Feature1"],
+            grid_figsize=(5, 5),  # This must be passed
+        )
+
+
+def test_kde_distributions_invalid_vars_of_interest(sample_dataframe):
+    # Expect function to return early without plotting when invalid vars are passed
+    kde_distributions(
+        df=sample_dataframe,
+        vars_of_interest=["NonExistentFeature"],
+    )
+    # No exception or warning expected per function design
+    assert True  # Smoke test
+
+
+def test_kde_distributions_log_scale_invalid_column(sample_dataframe):
+    with pytest.raises(ValueError, match="Invalid log_scale_vars"):
+        kde_distributions(
+            df=sample_dataframe,
+            vars_of_interest=["Feature1"],
+            log_scale_vars=["NonExistentFeature"],
+            plot_type="hist",
+            show_legend=False,
+        )
+
+
+def test_kde_distributions_invalid_y_axis(sample_dataframe):
+    with pytest.raises(ValueError, match="Invalid stat value"):
+        kde_distributions(
+            df=sample_dataframe,
+            vars_of_interest=["Feature1"],
+            stat="foobar",  # Invalid
+        )
+
+
+def test_kde_distributions_show_legend_no_labels(sample_dataframe):
+    kde_distributions(
+        df=sample_dataframe,
+        vars_of_interest=["Feature1"],
+        plot_mean=False,
+        plot_median=False,
+        std_dev_levels=[],
+        show_legend=True,
+    )
+
+
+def test_kde_distributions_tight_layout_error_handled(sample_dataframe, mocker):
+    mocker.patch(
+        "matplotlib.pyplot.tight_layout",
+        side_effect=RuntimeError("test error"),
+    )
+
+    with pytest.raises(RuntimeError, match="test error"):
+        kde_distributions(
+            df=sample_dataframe,
+            vars_of_interest=["Feature1"],
+        )
+
+
+def test_kde_distributions_with_legend_title(sample_scatter_dataframe):
+    kde_distributions(
+        df=sample_scatter_dataframe,
+        vars_of_interest=["Feature1"],
+        hue="Category",
+        show_legend=True,
+        legend_title="Test Legend",
+    )
+
+
+def test_kde_distributions_mismatched_legend_labels(sample_dataframe):
+    # legend_labels_list isn't a real param in kde_distributions, so remove test
+    # or stub it as smoke
+    kde_distributions(
+        df=sample_dataframe,
+        vars_of_interest=["Feature1"],
+        std_dev_levels=[1, 2],
+        std_color=["#888", "#999"],
+        show_legend=True,
+    )
+    assert True  # Stubbed until legend label logic is added
+
+
+def test_kde_distributions_no_save_without_filename(
+    sample_dataframe,
+    tmp_path,
+):
+    kde_distributions(
+        df=sample_dataframe,
+        vars_of_interest=["Feature1"],
+        image_path_png=str(tmp_path),
+        image_filename=None,  # Should not trigger saving
+    )
+    saved_files = list(tmp_path.glob("*.png"))
+    assert (
+        len(saved_files) == 0
+    ), "No files should be saved when image_filename is None."
+
+
+def test_kde_distributions_adds_legend(sample_dataframe):
+    kde_distributions(
+        df=sample_dataframe,
+        vars_of_interest=["Feature1"],
+        plot_mean=True,
+        plot_median=True,
+        std_dev_levels=[1],
+        show_legend=True,
+    )
+    # Cannot assert legend presence without UI, but should run without error
 
 
 def test_data_doctor_basic(sample_dataframe_values):
@@ -1417,138 +1354,3 @@ def test_stacked_crosstab_plot_save_multiple_formats(
 
     except Exception as e:
         pytest.fail(f"stacked_crosstab_plot failed with multiple save formats: {e}")
-
-
-def test_datetime_import_version_gte_3_7(mocker):
-    """Test datetime import behavior for Python versions >= 3.7"""
-    mocker.patch.object(sys, "version_info", (3, 7))  # Mock Python 3.7+
-
-    # Import datetime AFTER patching to ensure it behaves as expected
-    from datetime import datetime
-
-    assert "datetime" in sys.modules, "datetime module should be imported"
-    assert hasattr(
-        datetime, "fromisoformat"
-    ), "datetime should support 'fromisoformat' in Python 3.7+"
-
-
-def test_datetime_import_version_lt_3_7():
-    """Test datetime import behavior for Python versions < 3.7."""
-
-    # Mock `sys.version_info` to simulate Python 3.6
-    with patch.object(sys, "version_info", (3, 6)):
-
-        # Create a MagicMock instance for datetime.datetime
-        mock_datetime = MagicMock()
-        del (
-            mock_datetime.fromisoformat
-        )  # Remove `fromisoformat` to simulate Python <3.7
-
-        # Patch `datetime.datetime` with our mocked version
-        with patch("datetime.datetime", new=mock_datetime):
-            assert not hasattr(
-                datetime.datetime, "fromisoformat"
-            ), "datetime should NOT support 'fromisoformat' in Python <3.7"
-
-
-def test_ensure_directory_creates_new(mocker, tmp_path):
-    """
-    Test ensure_directory creates the directory when it does not exist.
-    """
-    dir_path = tmp_path / "new_dir"
-
-    # Mock os.path.exists to return False initially
-    mocker.patch("os.path.exists", return_value=False)
-    mock_makedirs = mocker.patch("os.makedirs")
-
-    ensure_directory(str(dir_path))
-
-    # Ensure makedirs was called since the directory didn’t exist
-    mock_makedirs.assert_called_once_with(str(dir_path))
-
-
-def test_ensure_directory_already_exists(mocker, tmp_path):
-    """
-    Test ensure_directory when directory already exists.
-    """
-    dir_path = tmp_path / "existing_dir"
-
-    # Mock os.path.exists to return True
-    mocker.patch("os.path.exists", return_value=True)
-    mock_makedirs = mocker.patch("os.makedirs")
-
-    ensure_directory(str(dir_path))
-
-    # Ensure makedirs was NOT called since the directory exists
-    mock_makedirs.assert_not_called()
-
-
-def test_highlight_columns():
-    # Create a sample DataFrame
-    data = {"A": [1, 2], "B": [3, 4], "C": [5, 6]}
-    df = pd.DataFrame(data)
-
-    # Apply highlight_columns function
-    styled_df = highlight_columns(df, columns=["A", "C"], color="lightblue")
-
-    # Ensure the return type is a pandas Styler
-    assert isinstance(
-        styled_df, pd.io.formats.style.Styler
-    ), "Function should return a Styler object"
-
-    # Extract the applied styles
-    styles = styled_df._compute().ctx  # Extracts computed styles as a dictionary
-
-    # Convert column names to positions
-    col_positions = {col: pos for pos, col in enumerate(df.columns)}
-
-    # Verify the expected styles
-    expected_styles = {
-        (0, col_positions["A"]): [("background-color", "lightblue")],
-        (1, col_positions["A"]): [("background-color", "lightblue")],
-        (0, col_positions["C"]): [("background-color", "lightblue")],
-        (1, col_positions["C"]): [("background-color", "lightblue")],
-    }
-
-    for key, value in expected_styles.items():
-        assert (
-            styles.get(key) == value
-        ), f"Expected {value} at {key}, but got {styles.get(key)}"
-
-    # Ensure column "B" remains unstyled
-    assert all(
-        styles.get((i, col_positions["B"]), []) == [] for i in range(len(df))
-    ), "Column B should not be styled"
-
-
-def test_custom_help_module(capsys):
-    """Test that custom_help() prints ASCII art and documentation when called with None."""
-    custom_help()
-    captured = capsys.readouterr()
-    assert eda_toolkit_logo in captured.out, "ASCII art should be printed."
-    assert detailed_doc in captured.out, "Detailed documentation should be printed."
-
-
-def test_custom_help_other_objects(capsys):
-    """Test that custom_help(obj) calls the original help function for other objects."""
-    obj = int  # Using a built-in type as a test case
-    original_help = builtins.help  # Capture the original help function
-
-    custom_help(obj)
-    captured = capsys.readouterr()
-
-    # Verify that ASCII art and custom docs are NOT printed
-    assert eda_toolkit_logo not in captured.out
-    assert detailed_doc not in captured.out
-
-    # Ensure original help() output is present
-    assert (
-        "class int" in captured.out or "Help on class int" in captured.out
-    ), "Original help should be invoked for int."
-
-
-def test_custom_help_override():
-    """Test that builtins.help is overridden by custom_help."""
-    assert (
-        builtins.help == custom_help
-    ), "builtins.help should be overridden by custom_help."
